user main template

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -15,7 +15,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        $categories = Category::paginate(2);
+        $categories = Category::paginate(5);
 
         return view('admin.categories.index', compact('categories'));
     }
@@ -91,8 +91,12 @@ class CategoryController extends Controller
     public function destroy($id)
     {
         $category = Category::find($id);
+
+        if ($category->posts->count()) {
+            return redirect()->route('categories.index')->with('error', 'Category has posts!');
+        }
+
         $category->delete();
-//        Category::destroy($id);
         return redirect()->route('categories.index')->with('success', 'Category has been successfully deleted');
     }
 }
